Move when_policy to util
